# job-portal-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Job Portal REST API",
    "description": "Comprehensive REST API documentation for the Job Portal system supporting Candidates, Companies, Jobs, Applications, Messaging, and Shortlisting.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occurred."
          }
        }
      },
...
      "CandidateProfileUpdate": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string", "example": "Jatin" },
          "lastName": { "type": "string", "example": "Verma" },
          "mobileNo": { "type": "string", "example": "+1234567890" },
          "bio": { "type": "string", "example": "Experienced software developer." },
          "dob": { "type": "string", "format": "date", "example": "1998-05-12" },
          "visibility": { "type": "string", "enum": ["public", "hidden", "private"], "example": "public" },
          "address": {
            "type": "object",
            "properties": {
              "street": { "type": "string", "example": "123 Main St" },
              "city": { "type": "string", "example": "Jaipur" },
              "state": { "type": "string", "example": "Rajasthan" },
              "country": { "type": "string", "example": "India" },
              "postalCode": { "type": "string", "example": "302001" }
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "skill": { "type": "string", "example": "React" },
                "level": { "type": "string", "example": "Advanced" }
              }
            }
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "language": { "type": "string", "example": "English" },
                "level": { "type": "string", "example": "Fluent" }
              }
            }
          },
          "education": { "type": "array", "items": { "type": "object" } },
          "experience": { "type": "array", "items": { "type": "object" } },
          "certifications": { "type": "array", "items": { "type": "object" } },
          "desiredJob": { "type": "object" }
        }
      },
      "CompanyProfileUpdate": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string", "example": "Aman" },
          "lastName": { "type": "string", "example": "Don" },
          "mobileNo": { "type": "string", "example": "+1234567890" },
          "bio": { "type": "string", "example": "Company Administrator" },
          "name": { "type": "string", "example": "Acme Corp" },
          "taxId": { "type": "string", "example": "TAX123" },
          "vatNo": { "type": "string", "example": "VAT123" },
          "industryId": { "type": "integer", "example": 1 },
          "size": { "type": "string", "example": "50-100" },
          "website": { "type": "string", "example": "https://acme.org" },
          "logo": { "type": "string", "example": "https://acme.org/logo.png" },
          "description": { "type": "string", "example": "Industry leading software providers." },
          "contactPerson": { "type": "string", "example": "Aman" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/candidates/{id}": {
      "get": {
        "tags": ["Candidates"],
        "summary": "Get candidate profile",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "7bc31252-bb79-457b-a040-6191e16f39b6"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Candidate not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      },
      "put": {
        "tags": ["Candidates"],
        "summary": "Update candidate profile (consolidated endpoint)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "7bc31252-bb79-457b-a040-6191e16f39b6"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Profile updated successfully" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized to update this profile"
          }
        }
      }
    },
    "/api/candidates/{id}/applications": {
      "get": {
        "tags": ["Candidates"],
        "summary": "Get candidate's job applications",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "7bc31252-bb79-457b-a040-6191e16f39b6"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{id}": {
      "get": {
        "tags": ["Companies"],
        "summary": "Get company profile",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "66666666-6666-6666-6666-666666666667"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Companies"],
        "summary": "Update company profile",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "66666666-6666-6666-6666-666666666667"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Company profile updated successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}